requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cmdop==2026.3.5.1
cryptography==46.0.5
distro==1.9.0
fastapi==0.135.1
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-genai==1.66.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
linkify-it-py==2.1.0
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
openclaw==2026.3.3
platformdirs==4.9.4
proto-plus==1.27.1
protobuf==5.29.6
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pyte==0.8.2
python-dotenv==1.2.2
python-telegram-bot==22.6
requests==2.32.5
rich==14.3.3
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.52.1
tenacity==9.1.4
textual==8.0.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uc-micro-py==2.0.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
wcwidth==0.6.0
websockets==16.0
